+++ JOA_Paris/app/src/main/java/com/sunah/foreignparis/korea/places/PlaceDetailFragment.kt
@@ -1,8 +1,20 @@
 package com.sunah.foreignparis.korea.places
 
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.MarkerOptions
+import com.sunah.foreignparis.korea.R
 import com.sunah.foreignparis.korea.places.model.PlaceInfoModel
+import kotlinx.android.synthetic.main.fragment_place.*
+import kotlinx.android.synthetic.main.fragment_place.view.*
+import kotlinx.android.synthetic.main.fragment_place_detail.*
+import kotlinx.android.synthetic.main.list_item_view.view.*
 
 class PlaceDetailFragment: Fragment() {
 
@@ -14,5 +26,19 @@ class PlaceDetailFragment: Fragment() {
             place = it.getParcelable("Details")!!
             val placename : String = place.address
         }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_place_detail, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+       list_item.info_name
+        
     }
 }
