+++ JOA_Paris/app/src/main/res/layout/fragment_place_detail.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical">
+
+            <ImageView
+                android:id="@+id/iv_"
+                android:layout_width="match_parent"
+                android:layout_height="168dp"
+                android:src="@drawable/ic_launcher_background"
+                app:layout_constraintTop_toTopOf="parent"
+                tools:layout_editor_absoluteX="122dp"
+                tools:layout_editor_absoluteY="43dp"
+                tools:src="@tools:sample/avatars"/>
+
+            <com.sunah.foreignparis.korea.places.ListItemView
+                android:id="@+id/list_item"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                android:layout_marginTop="16dp" />
+        </LinearLayout>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
