+++ JOA_Paris/app/src/main/res/layout/fragment_place_list.xml
