--- JOA_Paris/app/src/main/res/layout/using_custom_item_view.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".places.ListItemFragment">
-
-    <ImageView
-        android:id="@+id/iv_place"
-        android:layout_width="match_parent"
-        android:layout_height="168dp"
-        android:src="@drawable/ic_launcher_background"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="122dp"
-        tools:layout_editor_absoluteY="43dp"
-        tools:src="@tools:sample/avatars"/>
-
-
-        <com.sunah.foreignparis.korea.places.ListItemView
-            android:id="@+id/list_item"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:layout_marginTop="16dp"
-            app:layout_constraintTop_toBottomOf="@id/iv_place"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
