+++ JOA_Paris/app/src/main/res/values/attrs.xml
@@ -2,16 +2,6 @@
 <resources>
     <declare-styleable name="ListItemView">
         <attr name="image" format="reference"/>
-        <attr name="textSize" format="dimension"/>
-        <attr name="textColor" format="color|reference"/>
-        <attr name="secondaryTextColor" format="color|reference"/>
         <attr name="text" format="string"/>
-        <attr name="textPercentage" format="integer"/>
-        <attr name="suffixSymbol" format="string"/>
-        <attr name="textStyle">
-            <enum name="normal" value="0"/>
-            <enum name="bold" value="1"/>
-            <enum name="italic" value="2"/>
-        </attr>
     </declare-styleable>
 </resources>
